requirements.txt
function_frameworks
requests
pandas
bigquery
python-dotenv